fix: Recent visit handle-click fucntion fixed.
