+++ frontend/src/ui/NavBar.jsx
@@ -349,20 +349,7 @@ function NavBar() {
   // console.log("recentPlaylists", recentPlaylists);
   const navigate = useNavigate();
 
-  const handlePostClick = (post) => {
-    setSearchTerm("");
-    setInputValue("");
-    navigate(`/viewpage/${post.authorEmail}/${post._id}`);
-    setOpen(false);
-  };
-
-  const handlePlaylistClick = (playlist) => {
-    setSearchTerm("");
-    setInputValue("");
-
-    navigate(`/viewplaylist/${playlist._id}`);
-    setOpen(false);
-  };
+ 
 
   return (
     <div
